refactor: Move is_filepath_excluded to ggshield.core.file_utils
This function is a file related util.

refs #521
