+++ ggshield/core/file_utils.py
@@ -2,14 +2,17 @@ import os
 import re
 import tarfile
 from io import BytesIO
-from pathlib import Path
+from pathlib import Path, PurePosixPath
 from typing import List, Set, Union
 
 import click
 
-from ggshield.core.filter import is_filepath_excluded
 from ggshield.core.git_shell import git_ls, is_git_dir
 from ggshield.utils._binary_extensions import BINARY_EXTENSIONS
+
+
+def is_filepath_excluded(filepath: str, exclusion_regexes: Set[re.Pattern]) -> bool:
+    return any(r.search(str(PurePosixPath(Path(filepath)))) for r in exclusion_regexes)
 
 
 def get_filepaths(
