+++ tests/unit/core/test_filter.py
@@ -1,8 +1,6 @@
 import copy
 import random
-import re
-from pathlib import Path
-from typing import Iterable, List, Set, Union
+from typing import Iterable, List, Set
 
 import pytest
 from pygitguardian.models import Match, PolicyBreak, ScanResult
@@ -12,7 +10,6 @@ from ggshield.core.filter import (
     censor_content,
     censor_match,
     get_ignore_sha,
-    is_filepath_excluded,
     remove_ignored_from_result,
 )
 from ggshield.core.types import IgnoredMatch
@@ -240,18 +237,3 @@ def test_censor_content(content: str, policy_breaks: List[PolicyBreak]) -> None:
     for policy_break in policy_breaks:
         for match in policy_break.matches:
             assert match.match not in new_content
-
-
-@pytest.mark.parametrize(
-    "path,regexes,excluded",
-    [
-        ("foo", {"foo"}, True),
-        (Path("dir/foo"), {"foo"}, True),
-    ],
-)
-def test_is_filepath_excluded(
-    path: Union[str, Path], regexes: Set[str], excluded: bool
-) -> None:
-    path = str(path)
-    regexes = {re.compile(x) for x in regexes}
-    assert is_filepath_excluded(path, regexes) == excluded
